Moved ModelInputs into own module
